fix(licensure): break a module-init cycle, and say what is actually pending
Two changes, and the first is only here because the second surfaced it.

THE CYCLE

  sourceRouter → imports getRuntimeState from runtimeState
  runtimeState → imported isPhysicianProfession from sourceRouter

On any import order that initialised `runtimeState` first, its module body
called `isPhysicianProfession` while `sourceRouter`'s
`PHYSICIAN_PROFESSIONS` const was still in the temporal dead zone:

  ReferenceError: Cannot access 'PHYSICIAN_PROFESSIONS' before initialization

Minified, that reads `Cannot access 'g' before initialization`, which looks
like a bundler fault rather than a cycle in our own package — and that
exact string was already appearing in the /trust server log.

The cycle was latent: it only bit on orders reaching `runtimeState` first,
so most callers never saw it. Importing `coverageLabel` into a web
component was one such order, and it took the page down.

The predicates are pure data with no dependencies, so they move to a leaf
module that cannot participate in a cycle at all. `sourceRouter` imports
them for local use AND re-exports them, so every existing import path
keeps working. The plain import matters: a bare `export … from` does not
bind the names in the module's own scope, and `tsc` caught that after the
vitest suite had already gone green — the build is the stricter gate.

THE LABEL

The licensure lane read `Access required`, which is true of every gated
lane and therefore says nothing about this one. What is pending is access
to a NATIONAL network, so the state word now names that.

The phrase comes from the licensure package's own projection rather than
being typed into the component, because that projection is
self-correcting: `coverageLabel` derives the suffix from
`countLiveRoutes()` and stops saying "pending" the moment a route
completes a real production run. A literal here would keep saying pending
forever, including on the day it became false — the exact failure this
wave exists to remove.

Only the SUFFIX is used. `coverageLabel` returns a profession-scoped
headline ("Physician licensure — …") and this register is
profession-agnostic: it describes VitalCV's lanes, not one clinician's.
Rendering "Physician licensure" would silently drop nursing from a diagram
that covers it.

Both render sites change together. The aria-label uses the same helper as
the visible text, because a screen-reader user hearing "Licensure: access
required" while the diagram reads "national source access pending" is
being told a different, vaguer fact.

Verified on a real HTTP 200 (122kB body), not a grep against an error page:

  /trust   "State board"                    prod ×4  →  ×0
           "National source access pending"          →  ×3
           visible text and aria-label agree

Web: 385 files, 3547 tests. Licensure package suite green. Build exit 0.
Gates: typecheck, check:design, check:claims, check:routes, lint all 0.
